requirements.txt
apache-airflow==3.1.6
psycopg2-binary==2.9.11
